@@ -26,7 +26,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
 	"net/url"
@@ -39,6 +38,8 @@ import (
 	"sync"
 	"testing"
 	"time"
+
+	"golang.org/x/exp/slices"
 
 	"github.com/aosedge/aos_common/aoserrors"
 	"github.com/aosedge/aos_common/aostypes"
@@ -161,15 +162,17 @@ func TestInstallService(t *testing.T) {
 
 	cases := []struct {
 		serviceID            string
+		removeServiceID      string
 		version              string
 		size                 uint64
 		expectedCountService int
 		installErr           error
 		serviceConfig        aostypes.ServiceConfig
 	}{
+		// Install first service
 		{
 			serviceID:            "service1",
-			version:              "1.0",
+			version:              "1.0.0",
 			size:                 1 * megabyte,
 			expectedCountService: 1,
 			installErr:           nil,
@@ -182,9 +185,10 @@ func TestInstallService(t *testing.T) {
 				Resources: []string{"resource1", "resource2"},
 			},
 		},
+		// Install second service
 		{
 			serviceID:            "service2",
-			version:              "1.0",
+			version:              "1.0.0",
 			size:                 1 * megabyte,
 			expectedCountService: 2,
 			installErr:           nil,
@@ -197,9 +201,42 @@ func TestInstallService(t *testing.T) {
 				Resources: []string{"resource1"},
 			},
 		},
+		// Install same version service
+		{
+			serviceID:            "service1",
+			version:              "1.0.0",
+			size:                 1 * megabyte,
+			expectedCountService: 2,
+			installErr:           nil,
+			serviceConfig: aostypes.ServiceConfig{
+				Hostname: allocateString("service1"),
+				Quotas: aostypes.ServiceQuotas{
+					UploadSpeed:   allocateUint64(1000),
+					DownloadSpeed: allocateUint64(2000),
+				},
+				Resources: []string{"resource1", "resource2"},
+			},
+		},
+		// Install old version service on active service
+		{
+			serviceID:            "service1",
+			version:              "0.9.0",
+			size:                 1 * megabyte,
+			expectedCountService: 2,
+			installErr:           imagemanager.ErrVersionMismatch,
+		},
+		// Install old version service on cached service
+		{
+			serviceID:            "service1",
+			removeServiceID:      "service1",
+			version:              "0.9.0",
+			size:                 1 * megabyte,
+			expectedCountService: 2,
+		},
+		// Install other service with no space available
 		{
 			serviceID:            "service3",
-			version:              "1.0",
+			version:              "1.0.0",
 			size:                 2 * megabyte,
 			expectedCountService: 2,
 			installErr:           spaceallocator.ErrNoSpace,
@@ -219,7 +256,15 @@ func TestInstallService(t *testing.T) {
 		}
 	}()
 
-	for _, tCase := range cases {
+	for i, tCase := range cases {
+		t.Logf("Test case: %d", i)
+
+		if tCase.removeServiceID != "" {
+			if err := imagemanagerInstance.RemoveService(tCase.removeServiceID); err != nil {
+				t.Fatalf("Can't remove service: %v", err)
+			}
+		}
+
 		configJSON, err := json.Marshal(tCase.serviceConfig)
 		if err != nil {
 			t.Errorf("Can't generate config json: %v", err)
@@ -301,7 +346,7 @@ func TestRevertService(t *testing.T) {
 	}{
 		{
 			serviceID: "service1",
-			version:   "1.0",
+			version:   "1.0.0",
 			size:      1 * megabyte,
 			serviceConfig: aostypes.ServiceConfig{
 				Hostname: allocateString("service1"),
@@ -314,7 +359,20 @@ func TestRevertService(t *testing.T) {
 		},
 		{
 			serviceID: "service1",
-			version:   "2.0",
+			version:   "2.0.0",
+			size:      1 * megabyte,
+			serviceConfig: aostypes.ServiceConfig{
+				Hostname: allocateString("service1"),
+				Quotas: aostypes.ServiceQuotas{
+					UploadSpeed:   allocateUint64(500),
+					DownloadSpeed: allocateUint64(1000),
+				},
+				Resources: []string{"resource1"},
+			},
+		},
+		{
+			serviceID: "service1",
+			version:   "3.0.0",
 			size:      1 * megabyte,
 			serviceConfig: aostypes.ServiceConfig{
 				Hostname: allocateString("service1"),
@@ -355,22 +413,25 @@ func TestRevertService(t *testing.T) {
 	}{
 		{
 			revertServiceID: "service1",
-			expectedVersion: "1.0",
+			expectedVersion: "2.0.0",
 			err:             nil,
 		},
 		{
 			revertServiceID: "service1",
 			err:             imagemanager.ErrNotExist,
+			expectedVersion: "2.0.0",
 		},
 	}
 
-	for _, tCase := range casesRevert {
-		if err := imagemanagerInstance.RevertService(tCase.revertServiceID); err != nil {
+	for i, tCase := range casesRevert {
+		t.Logf("Test case: %d", i)
+
+		if err := imagemanagerInstance.RevertService(tCase.revertServiceID); !errors.Is(err, tCase.err) {
 			t.Errorf("Can't revert service to previous version: %v", err)
 		}
 
 		service, err := imagemanagerInstance.GetServiceInfo(tCase.revertServiceID)
-		if !errors.Is(err, tCase.err) {
+		if err != nil {
 			t.Errorf("Can't get service: %v", err)
 		}
 
@@ -418,7 +479,7 @@ func TestRemoveService(t *testing.T) {
 	}{
 		{
 			serviceID:            "service1",
-			version:              "1.0",
+			version:              "1.0.0",
 			size:                 1 * megabyte,
 			expectedCountService: 1,
 			installErr:           nil,
@@ -434,7 +495,7 @@ func TestRemoveService(t *testing.T) {
 		{
 			serviceID:            "service2",
 			removeServiceID:      "service1",
-			version:              "1.0",
+			version:              "1.0.0",
 			size:                 1 * megabyte,
 			expectedCountService: 1,
 			installErr:           spaceallocator.ErrNoSpace,
@@ -449,7 +510,9 @@ func TestRemoveService(t *testing.T) {
 		},
 	}
 
-	for _, tCase := range cases {
+	for i, tCase := range cases {
+		t.Logf("Test case: %d", i)
+
 		configJSON, err := json.Marshal(tCase.serviceConfig)
 		if err != nil {
 			t.Errorf("Can't generate config json: %v", err)
@@ -474,13 +537,8 @@ func TestRemoveService(t *testing.T) {
 				t.Errorf("Can't remove service: %v", err)
 			}
 
-			service, err := imagemanagerInstance.GetServiceInfo(tCase.removeServiceID)
-			if err != nil {
-				t.Errorf("Can't get service info: %v", err)
-			}
-
-			if !service.Cached {
-				t.Error("Service should be cached")
+			if _, err := imagemanagerInstance.GetServiceInfo(tCase.removeServiceID); err == nil {
+				t.Error("Service should be removed")
 			}
 
 			if err := imagemanagerInstance.InstallService(serviceInfo, nil, nil); err != nil {
@@ -550,7 +608,7 @@ func TestRestoreService(t *testing.T) {
 	}{
 		{
 			serviceID:  "service1",
-			version:    "1.0",
+			version:    "1.0.0",
 			size:       1 * megabyte,
 			installErr: nil,
 			serviceConfig: aostypes.ServiceConfig{
@@ -565,7 +623,7 @@ func TestRestoreService(t *testing.T) {
 		{
 			serviceID:       "service2",
 			removeServiceID: "service1",
-			version:         "1.0",
+			version:         "1.0.0",
 			size:            1 * megabyte,
 			installErr:      spaceallocator.ErrNoSpace,
 			serviceConfig: aostypes.ServiceConfig{
@@ -604,24 +662,20 @@ func TestRestoreService(t *testing.T) {
 				t.Errorf("Can't remove service: %v", err)
 			}
 
-			service, err := imagemanagerInstance.GetServiceInfo(tCase.removeServiceID)
-			if err != nil {
-				t.Errorf("Can't get service info: %v", err)
-			}
-
-			if !service.Cached {
-				t.Error("Service should be cached")
+			if _, err := imagemanagerInstance.GetServiceInfo(tCase.removeServiceID); err == nil {
+				t.Error("Service should be removed")
 			}
 
 			if err := imagemanagerInstance.RestoreService(tCase.removeServiceID); err != nil {
 				t.Errorf("Can't restore service: %v", err)
 			}
 
-			if service, err = imagemanagerInstance.GetServiceInfo(tCase.removeServiceID); err != nil {
+			service, err := imagemanagerInstance.GetServiceInfo(tCase.removeServiceID)
+			if err != nil {
 				t.Errorf("Can't get service info: %v", err)
 			}
 
-			if service.Cached {
+			if service.State == imagemanager.ServiceCached {
 				t.Error("Service should not be cached")
 			}
 
@@ -659,42 +713,50 @@ func TestRestoreLayer(t *testing.T) {
 	defer imagemanagerInstance.Close()
 
 	cases := []struct {
+		id                 string
+		version            string
 		digest             string
-		removedDigest      string
+		removeDigest       string
 		size               uint64
 		expectedCountLayer int
 		installErr         error
 	}{
 		{
+			id:                 "layer1",
+			version:            "1.0.0",
 			digest:             "digest1",
 			size:               1 * megabyte,
 			expectedCountLayer: 1,
 			installErr:         nil,
 		},
 		{
+			id:                 "layer2",
+			version:            "1.0.0",
 			digest:             "digest2",
 			size:               1 * megabyte,
 			expectedCountLayer: 1,
 			installErr:         nil,
 		},
 		{
+			id:                 "layer3",
+			version:            "1.0.0",
 			digest:             "digest3",
 			size:               1 * megabyte,
-			removedDigest:      "digest1",
+			removeDigest:       "digest1",
 			expectedCountLayer: 1,
 			installErr:         spaceallocator.ErrNoSpace,
 		},
 	}
 
-	for index, tCase := range cases {
-		fileName := path.Join(tmpDir, fmt.Sprintf("layer_%d", index))
+	for _, tCase := range cases {
+		fileName := path.Join(tmpDir, tCase.id)
 
 		if err = generateFile(fileName, tCase.size); err != nil {
 			t.Errorf("Can't generate file: %v", err)
 		}
 		defer os.RemoveAll(fileName)
 
-		layerInfo, err := prepareLayerInfo(fileName, tCase.digest, index)
+		layerInfo, err := prepareLayerInfo(fileName, tCase.id, tCase.version, tCase.digest)
 		if err != nil {
 			t.Errorf("Can't prepare layer info data: %v", err)
 		}
@@ -704,32 +766,28 @@ func TestRestoreLayer(t *testing.T) {
 		}
 
 		if tCase.installErr != nil {
-			if err := imagemanagerInstance.RemoveLayer(tCase.removedDigest); err != nil {
+			if err := imagemanagerInstance.RemoveLayer(tCase.removeDigest); err != nil {
 				t.Errorf("Can't remove layer: %v", err)
 			}
 
-			layer, err := imagemanagerInstance.GetLayerInfo(tCase.removedDigest)
+			if _, err := imagemanagerInstance.GetLayerInfo(tCase.removeDigest); err == nil {
+				t.Error("Layer should be removed")
+			}
+
+			if err := imagemanagerInstance.RestoreLayer(tCase.removeDigest); err != nil {
+				t.Errorf("Can't restore layer: %v", err)
+			}
+
+			layer, err := imagemanagerInstance.GetLayerInfo(tCase.removeDigest)
 			if err != nil {
 				t.Errorf("Can't get layer info: %v", err)
 			}
 
-			if !layer.Cached {
-				t.Error("Layer should be cached")
-			}
-
-			if err := imagemanagerInstance.RestoreLayer(tCase.removedDigest); err != nil {
-				t.Errorf("Can't restore layer: %v", err)
-			}
-
-			if layer, err = imagemanagerInstance.GetLayerInfo(tCase.removedDigest); err != nil {
-				t.Errorf("Can't get layer info: %v", err)
-			}
-
-			if layer.Cached {
+			if layer.State == imagemanager.LayerCached {
 				t.Error("Layer should not be cached")
 			}
 
-			if err := imagemanagerInstance.RemoveLayer(tCase.removedDigest); err != nil {
+			if err := imagemanagerInstance.RemoveLayer(tCase.removeDigest); err != nil {
 				t.Errorf("Can't remove layer: %v", err)
 			}
 
@@ -737,7 +795,7 @@ func TestRestoreLayer(t *testing.T) {
 				t.Errorf("Can't install layer: %v", err)
 			}
 
-			if err := imagemanagerInstance.RestoreLayer(tCase.removedDigest); err == nil {
+			if err := imagemanagerInstance.RestoreLayer(tCase.removeDigest); err == nil {
 				t.Error("Expected error layer not found")
 			}
 		}
@@ -763,6 +821,8 @@ func TestRemoveLayer(t *testing.T) {
 	defer imagemanagerInstance.Close()
 
 	cases := []struct {
+		id                 string
+		version            string
 		digest             string
 		removeDigest       string
 		size               uint64
@@ -770,12 +830,16 @@ func TestRemoveLayer(t *testing.T) {
 		installErr         error
 	}{
 		{
+			id:                 "layer1",
+			version:            "1.0.0",
 			digest:             "digest1",
 			size:               1 * megabyte,
 			expectedCountLayer: 1,
 			installErr:         nil,
 		},
 		{
+			id:                 "layer2",
+			version:            "1.0.0",
 			digest:             "digest2",
 			removeDigest:       "digest1",
 			size:               2 * megabyte,
@@ -784,15 +848,15 @@ func TestRemoveLayer(t *testing.T) {
 		},
 	}
 
-	for index, tCase := range cases {
-		fileName := path.Join(tmpDir, fmt.Sprintf("layer_%d", index))
+	for _, tCase := range cases {
+		fileName := path.Join(tmpDir, tCase.id)
 
 		if err = generateFile(fileName, tCase.size); err != nil {
 			t.Errorf("Can't generate file: %v", err)
 		}
 		defer os.RemoveAll(fileName)
 
-		layerInfo, err := prepareLayerInfo(fileName, tCase.digest, index)
+		layerInfo, err := prepareLayerInfo(fileName, tCase.id, tCase.version, tCase.digest)
 		if err != nil {
 			t.Errorf("Can't prepare layer info data: %v", err)
 		}
@@ -806,13 +870,8 @@ func TestRemoveLayer(t *testing.T) {
 				t.Errorf("Can't remove layer: %v", err)
 			}
 
-			layer, err := imagemanagerInstance.GetLayerInfo(tCase.removeDigest)
-			if err != nil {
-				t.Errorf("Can't get layer info: %v", err)
-			}
-
-			if !layer.Cached {
-				t.Error("Layer should be cached")
+			if _, err := imagemanagerInstance.GetLayerInfo(tCase.removeDigest); err == nil {
+				t.Error("Layer should be removed")
 			}
 
 			if err := imagemanagerInstance.InstallLayer(layerInfo, nil, nil); err != nil {
@@ -864,30 +923,64 @@ func TestInstallLayer(t *testing.T) {
 	defer imagemanagerInstance.Close()
 
 	cases := []struct {
+		id                 string
+		version            string
 		digest             string
+		removeDigest       string
 		size               uint64
 		expectedCountLayer int
 		installErr         error
 	}{
+		// install first layer
 		{
+			id:                 "layer1",
+			version:            "1.0.0",
 			digest:             "digest1",
 			size:               1 * megabyte,
 			expectedCountLayer: 1,
 			installErr:         nil,
 		},
+		// install second layer
 		{
+			id:                 "layer2",
+			version:            "1.0.0",
 			digest:             "digest2",
 			size:               1 * megabyte,
 			expectedCountLayer: 2,
 			installErr:         nil,
 		},
+		// install same version
 		{
+			id:                 "layer1",
+			version:            "1.0.0",
 			digest:             "digest1",
 			size:               1 * megabyte,
 			expectedCountLayer: 2,
 			installErr:         nil,
 		},
+		// install other version
 		{
+			id:                 "layer1",
+			version:            "2.0.0",
+			digest:             "digest1",
+			size:               1 * megabyte,
+			expectedCountLayer: 2,
+			installErr:         nil,
+		},
+		// restore cached layer
+		{
+			id:                 "layer1",
+			version:            "2.0.0",
+			digest:             "digest1",
+			removeDigest:       "digest1",
+			size:               1 * megabyte,
+			expectedCountLayer: 2,
+			installErr:         nil,
+		},
+		// install layer with no enough space
+		{
+			id:                 "layer3",
+			version:            "1.0.0",
 			digest:             "digest3",
 			size:               2 * megabyte,
 			expectedCountLayer: 2,
@@ -901,30 +994,39 @@ func TestInstallLayer(t *testing.T) {
 		}
 	}()
 
-	for index, tCase := range cases {
-		fileName := path.Join(tmpDir, fmt.Sprintf("layer_%d", index))
+	for i, tCase := range cases {
+		t.Logf("Test case: %d", i)
+
+		fileName := path.Join(tmpDir, tCase.id)
 
 		if err = generateFile(fileName, tCase.size); err != nil {
 			t.Errorf("Can't generate file: %v", err)
 		}
 		defer os.RemoveAll(fileName)
 
-		layerInfo, err := prepareLayerInfo(fileName, tCase.digest, index)
+		layerInfo, err := prepareLayerInfo(fileName, tCase.id, tCase.version, tCase.digest)
 		if err != nil {
 			t.Errorf("Can't prepare layer info data: %v", err)
 		}
 
+		if tCase.removeDigest != "" {
+			if err := imagemanagerInstance.RemoveLayer(tCase.removeDigest); err != nil {
+				t.Fatalf("Can't remove layer: %v", err)
+			}
+		}
+
 		if err = imagemanagerInstance.InstallLayer(layerInfo, nil, nil); !errors.Is(err, tCase.installErr) {
-			t.Errorf("Can't install layer: %v", err)
+			t.Fatalf("Can't install layer: %v", err)
 		}
 
 		if tCase.installErr == nil {
 			layer, err := imagemanagerInstance.GetLayerInfo(tCase.digest)
 			if err != nil {
-				t.Errorf("Can't install layer: %v", err)
+				t.Errorf("Can't get layer info layer: %v", err)
 			}
 
-			if layer.Digest != tCase.digest || layer.Size != tCase.size {
+			if layer.LayerID != tCase.id || layer.Version != tCase.version || layer.Digest != tCase.digest ||
+				layer.Size != tCase.size || layer.State != imagemanager.LayerActive {
 				t.Error("Unexpected layer info")
 			}
 		}
@@ -935,7 +1037,7 @@ func TestInstallLayer(t *testing.T) {
 		}
 
 		if tCase.expectedCountLayer != len(layers) {
-			t.Error("Unexpected layers count")
+			t.Errorf("Unexpected layers count: %d", len(layers))
 		}
 
 		for _, layer := range layers {
@@ -973,7 +1075,7 @@ func TestFileServer(t *testing.T) {
 		t.Fatalf("Can't create package file: %s", err)
 	}
 
-	layerInfo, err := prepareLayerInfo(fileName, "digest", 1)
+	layerInfo, err := prepareLayerInfo(fileName, "layer1", "1.0.0", "digest")
 	if err != nil {
 		t.Fatalf("Can't prepare layer info data: %v", err)
 	}
@@ -1154,15 +1256,6 @@ func (storage *testStorageProvider) GetLayersInfo() (layers []imagemanager.Layer
 	return layers, nil
 }
 
-func (storage *testStorageProvider) GetServiceInfo(serviceID string) (imagemanager.ServiceInfo, error) {
-	services, ok := storage.services[serviceID]
-	if !ok {
-		return imagemanager.ServiceInfo{}, imagemanager.ErrNotExist
-	}
-
-	return services[len(services)-1], nil
-}
-
 func (storage *testStorageProvider) GetLayerInfo(digest string) (imagemanager.LayerInfo, error) {
 	layer, ok := storage.layers[digest]
 	if !ok {
@@ -1193,27 +1286,32 @@ func (storage *testStorageProvider) AddService(service imagemanager.ServiceInfo)
 	return nil
 }
 
-func (storage *testStorageProvider) SetLayerCached(digest string, cached bool) error {
+func (storage *testStorageProvider) SetLayerState(digest string, state int) error {
 	layer, ok := storage.layers[digest]
 	if !ok {
 		return aoserrors.New("layer not found")
 	}
 
-	layer.Cached = cached
+	layer.State = state
 	storage.layers[digest] = layer
 
 	return nil
 }
 
-func (storage *testStorageProvider) SetServiceCached(serviceID string, cached bool) error {
+func (storage *testStorageProvider) SetServiceState(serviceID, version string, state int) error {
 	services, ok := storage.services[serviceID]
 	if !ok {
 		return aoserrors.New("service not found")
 	}
 
-	for i := 0; i < len(services); i++ {
-		services[i].Cached = cached
+	index := slices.IndexFunc(services, func(service imagemanager.ServiceInfo) bool {
+		return service.Version == version
+	})
+	if index == -1 {
+		return imagemanager.ErrNotExist
 	}
+
+	services[index].State = state
 
 	return nil
 }
@@ -1255,7 +1353,7 @@ func (storage *testStorageProvider) RemoveLayer(digest string) error {
  * Private
  **********************************************************************************************************************/
 
-func prepareLayerInfo(filePath, digest string, index int) (layerInfo cloudprotocol.LayerInfo, err error) {
+func prepareLayerInfo(filePath, id, version, digest string) (layerInfo cloudprotocol.LayerInfo, err error) {
 	imageFileInfo, err := image.CreateFileInfo(context.Background(), filePath)
 	if err != nil {
 		return cloudprotocol.LayerInfo{}, nil
@@ -1275,8 +1373,8 @@ func prepareLayerInfo(filePath, digest string, index int) (layerInfo cloudprotoc
 	}
 
 	installRequest := cloudprotocol.LayerInfo{
-		Version: "v" + strconv.Itoa(index),
-		LayerID: "testLayer" + strconv.Itoa(index),
+		Version: version,
+		LayerID: id,
 		Digest:  digest,
 		DownloadInfo: cloudprotocol.DownloadInfo{
 			URLs:   []string{url.String()},
@@ -1405,8 +1503,6 @@ func prepareService(servicelayerSize uint64, srvConfig []byte,
 }
 
 func packImage(source, name string) (err error) {
-	log.WithFields(log.Fields{"source": source, "name": name}).Debug("Pack image")
-
 	if output, err := exec.Command("tar", "-C", source, "-cf", name, "./").CombinedOutput(); err != nil {
 		return aoserrors.Errorf("tar error: %s, code: %s", string(output), err)
 	}
